--- a/src/3rd_party/topk.cuh
+++ b/src/3rd_party/topk.cuh
@@ -25,10 +25,8 @@
 #include <cuda_fp16.h>
 #if CUDA_VERSION >= 11000
 #include <cub/cub.cuh>
-#include <cub/util_type.cuh>
 #else
 #include "cub/cub/cub.cuh"
-#include "cub/cub/util_type.cuh"
 #endif
 #define MAX_BLOCKS_PER_BEAM 8