--- a/src/graph/expression_operators.h
+++ b/src/graph/expression_operators.h
@@ -173,6 +173,8 @@ Expr atleast_3d(Expr a);
 Expr atleast_4d(Expr a);
 Expr atleast_nd(Expr a, size_t dims);
 
+Expr addPosEmbedding(Expr embeddings, float scaleFactor, int startPos);
+
 // create a constant of shape a->shape() and initialize with init
 // @TODO: add a && version, to avoid a ref count. NodeInitializers are typically temps.
 // @TODO: and/or make this a template on init