--- a/src/graph/expression_operators.h
+++ b/src/graph/expression_operators.h
@@ -1,3 +1,8 @@
+/* All or part of this file was contributed by NVIDIA under license:
+ *   Copyright (C) 2020 NVIDIA Corporation
+ *   SPDX-License-Identifier: MIT
+ */
+
 #pragma once
 #include "graph/expression_graph.h"
 #include "graph/node_initializers.h"