… CMakeLists.txt
… src/CMakeLists.txt
… src/layers/generic.h
… src/models/states.h
… src/models/transformer.h
… src/rnn/types.h
… src/tensors/cpu/prod.cpp
… src/tensors/cpu/tensor_operators.cpp
… src/tensors/dispatch.h
… src/tensors/gpu/algorithm.cu
… src/tensors/gpu/algorithm.h
… src/tensors/gpu/backend.h
… src/tensors/gpu/prod.cpp
… src/tensors/gpu/prod.cu
… src/tensors/gpu/prod.h
… src/tensors/gpu/tensor_operators.cu
… src/tensors/tensor.h
… src/tensors/tensor_operators.h
… src/translator/beam_search.cpp
… src/translator/beam_search.h
… src/translator/nth_element.cu