### Description

This PR represents a single branch containing all of the GPU optimizations done. It is mainly to get a wholistic view of the performance improvements from the main branch. It is being split into several smaller PRs and will be closed once the smaller PRs are merged into master.

Performance numbers from a proxy model. The times are the total time to decode an input file by splitting the file into different batch sizes. This was done since it is strange to compare the time per batch with multiple streams. All times were collected on a Titan V.

Times using one stream. 
Batch | Initial Time (s) | Current Time(s) | % Runtime reduction | Speedup factor
-- | -- | -- | -- | --
1 | 204.41 | 92.15 | 0.549190353 | 2.218231145
2 | 136.77 | 66.29 | 0.515317687 | 2.06320712
4 | 84.76 | 41.23 | 0.513567721 | 2.055784623
8 | 50.32 | 25.05 | 0.50218601 | 2.008782435
16 | 29.94 | 15 | 0.498997996 | 1.996
32 | 18.51 | 9.26 | 0.499729876 | 1.998920086
64 | 12.09 | 6.03 | 0.501240695 | 2.004975124
128 | 8.19 | 4.13 | 0.495726496 | 1.983050847
256 | 6.02 | 3.03 | 0.496677741 | 1.98679868



Times with two streams
Batch | Initial Time (s) | Current Time(s) | % Runtime reduction | Speedup factor
-- | -- | -- | -- | --
1 | 156.74 | 67.18 | 0.571392114 | 2.333134862
2 | 106.13 | 49.07 | 0.537642514 | 2.162828612
4 | 65.1 | 30.18 | 0.53640553 | 2.157057654
8 | 38.46 | 18.04 | 0.530941238 | 2.131929047
16 | 22.68 | 10.73 | 0.526895944 | 2.113699907
32 | 14.31 | 6.65 | 0.535290007 | 2.151879699
64 | 9.51 | 4.42 | 0.535226078 | 2.15158371
128 | 6.66 | 3.15 | 0.527027027 | 2.114285714
256 | 5.05 | 2.34 | 0.536633663 | 2.158119658



List of changes:
- Compute Sinusoidal Embeddings on GPU when compiled with CUDA instead of doing work on the CPU. This causes small float differences in the final result of some sentences.
- Reduces the number of times indices need to be transferred to the GPU when updating transformer and rnn states
- Issues cublas and cusparse calls to the per thread default stream instead of the global default stream
- Adds a operator to fuse bias addition with relu during inference. No longer issues a gemm to broadcast the bias across the tensor during inference.
- Caches the triangle mask in the transformer during inference to reduce communication
- Batches the retrieval of n-best logits from the device so one memcpy is issued to get all logits instead of 1 memcpy per logit
- Adds a new operator that fuses bias addition with the skip connection and layer normalization
- Removes memcpy from ProdBatched when the number of batches in the input matrices are the same by using cublasStridedBatchedGemm.
- Reduces the number of memcpys in ProdBatched from 3 to 1 in the general case
- Adds a cache by name function to the graph
- Moves the lemmaHasFactorGroup tensor to the GPU for factored vocabulary models to allow lemmas to be masked out on the GPU in one kernel for group 0.
- Ports the topk implementation from NVIDIA's faster transformer to nth_element.cu and make changes so that it is compatible with marian.

Added dependencies: 
1.  cub - used by the new topk implementation

### How to test
I ran the regression tests. They will fail due to some differences with float computations. Not sure of the protocol for updating the tests.

OS: Ubuntu 18.04.3 LTS
Compiler gcc 7.5.0
nvcc version: 10.1.243

cmake command:

cmake .. -DCOMPILE_CPU=on -DCOMPILE_CUDA=on -DUSE_SENTENCEPIECE=on -DUSE_STATIC_LIBS=off -DCOMPILE_SERVER=off -DUSE_FBGEMM=on -DCOMPILE_CUDA_SM35=off -DCOMPILE_CUDA_SM50=off -DCOMPILE_CUDA_SM60=off -DCOMPILE_CUDA_SM70=on

### Checklist

- [x] I have tested the code manually
- [x] I have run regression tests - Some regression tests fail from float differences
- [ ] I have read and followed CONTRIBUTING.md
- [ ] I have updated CHANGELOG.md