@rhenry-nv 2020-10-10
@ykim362 done! This is in draft status since there is another WIP optimization I plan to add for factored vocab models. End to end with preliminary testing on the WIP commit is 2.1x for batch 1 and 1.28x for batch 256 over the initial code using a proxy transformer model.

I also need to sort out the failing checks. I can compile locally but will look into replicating the exact build steps so I can fix the checks.