@rhenry-nv 2020-10-12
Yes, the same tests fail on my Titan V from the master branch. As a workaround, I increased the diff sensitivity in the python script to something like 0.2 to avoid spamming with failing regressions.

The output changes on my Titan V from commit 5b889da and shows up in the following regression tests (these are the ones I was referring to):

marian-dev/regression-tests/tests/models/transformer/test_nbest.sh                                                                                                                                                  marian-dev/regression-tests/tests/models/ape/test_nbest.sh
marian-dev/regression-tests/tests/training/validation/test_compare_decoding_with_transscript_output.sh.log

A small number of sentences have slight output differences.