@XapaJIaMnu 2020-10-20
I need `add_definitions(-DTHRUST_IGNORE_DEPRECATED_CPP_DIALECT=1)` In order to compile nth_element.cu. CUDA 11.0.