@XapaJIaMnu 2020-10-20
@kpu gcc 8.4.0
Also, Cub version redefinition:
```
In file included from /usr/local/cuda/include/cub/util_namespace.cuh:36,
[build]                  from /usr/local/cuda/include/thrust/system/cuda/config.h:76,
[build]                  from /usr/local/cuda/include/thrust/system/cuda/detail/execution_policy.h:33,
[build]                  from /usr/local/cuda/include/thrust/iterator/detail/device_system_tag.h:23,
[build]                  from /usr/local/cuda/include/thrust/iterator/iterator_traits.h:111,
[build]                  from /usr/local/cuda/include/thrust/system/cuda/detail/util.h:31,
[build]                  from /usr/local/cuda/include/thrust/system/cuda/detail/core/alignment.h:21,
[build]                  from /usr/local/cuda/include/thrust/system/cuda/detail/core/triple_chevron_launch.h:30,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/device/dispatch/dispatch_histogram.cuh:48,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/device/device_histogram.cuh:41,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/cub.cuh:52,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/topk.cuh:26,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/translator/nth_element.cu:9:
[build] /usr/local/cuda/include/cub/version.cuh:46: error: "CUB_VERSION" redefined [-Werror]
[build]  #define CUB_VERSION 100909
[build]  
[build] In file included from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/util_namespace.cuh:36,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/util_arch.cuh:37,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/config.cuh:35,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/cub.cuh:37,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/3rd_party/topk.cuh:26,
[build]                  from /home/s1031254/marian-dev-8bitgpu/src/translator/nth_element.cu:9:
[build] /home/s1031254/marian-dev-8bitgpu/src/3rd_party/cub/cub/version.cuh:46: note: this is the location of the previous definition
[build]  #define CUB_VERSION 101100
```