@rhenry-nv 2020-10-20
I have not compiled with CUDA 11 but the redefinition is because CUDA 11 now ships with CUB causing competing versions to exist. 

I was not aware that Marian compiled with CUDA 11. I will fix these issues in a future commit.