@XapaJIaMnu 2020-10-20
@rhenry-nv  we support up to CUDA 11.1, although not officially.