@XapaJIaMnu 2020-10-21
@rhenry-nv 
Here's a patch that fixes CUDA11+:
```patch
diff --git a/src/3rd_party/topk.cuh b/src/3rd_party/topk.cuh
index ad0e452..0df9a73 100644
--- a/src/3rd_party/topk.cuh
+++ b/src/3rd_party/topk.cuh
@@ -23,9 +23,13 @@
 #pragma once
 #include <cuda_runtime.h>
 #include <cuda_fp16.h>
+#if CUDA_VERSION >= 11000
+#include <cub/cub.cuh>
+#include <cub/util_type.cuh>
+#else
 #include "cub/cub/cub.cuh"
 #include "cub/cub/util_type.cuh"
-
+#endif
 #define MAX_BLOCKS_PER_BEAM 8
 
 struct TopK {
diff --git a/src/CMakeLists.txt b/src/CMakeLists.txt
index 33deb8a..7f08575 100644
--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -3,6 +3,7 @@
  #   SPDX-License-Identifier: MIT
 
 add_definitions(-DCUB_IGNORE_DEPRECATED_CPP_DIALECT=1)
+add_definitions(-DTHRUST_IGNORE_DEPRECATED_CPP_DIALECT=1)
 add_subdirectory(3rd_party)
 
 include_directories(.)
```