@emjotde 2020-11-04
BTW, the merge I mentioned above is in branch `mjd/fp16.3`.