@rhenry-nv 2020-11-04
> This is marked as a draft, so there is still active work going on, right?
> Due to that I was currently ignoring that, but I assume you want intermediate feedback?

Yes, the intention was to get intermediate feedback but I should've been clearer about that.

> @rhenry-nv Are you planning to keep adding new features to this PR, or should we try to get in what's here currently? There is so much stuff to review and test that I am a bit afraid it will overwhelm us a bit in a single PR.

I can try to split into separate PRs or just stop adding stuff to this PR. There are a few more changes that I have not added in here that I can make a separate PR for. Let me know what would be the best way to ensure that you're not overwhelmed.