@emjotde 2020-11-04
I am partial to splitting this up. I looked through it a bit, a there it at least a couple of separate "themes". I am seeing:
* small refactoring changes that give you a lot of bang for the buck, like: reusing index vectors, per-thread streams, and similar things. These could be one very non-controversial PR
* Fused operators. That also looks alright, especially when we manage to push the forks a bit further down the into the expression operators. 
* All kinds of other "hacks", like various kind of memoization and caching which seem to duplicate existing solutions or add more of something that we intended to actually get rid of. 
* Top-k operator and beam search also seems like a separate item for me, also we should unify with existing top-k stuff. 

I will spend some more time on finding my way through these changes and then set up a call if that is OK?
Anyway, thanks a lot for this, there is a lot of good changes in here at a first glance.